Factory refactored

+++ src/calc/abstractFactory/DivisionOperation.java
@@ -5,6 +5,6 @@ public class DivisionOperation implements IMathAction {
     @Override
     public double doAction(int x, int y) {
         System.out.print(x + " / " + y + " = ");
-        return x/y;
+        return (double) x/y;
     }
 }
